6964: Refactor - Unifying handling of Java version
